requirements.txt
scipy==1.13.1
numpy
scikit-learn==1.5.0
redis
torch
stable-baselines3
gym
shimmy
pynvml
python-dotenv
requests
